hardcode magician names
As a best practice it's better to have the magician names derived dynamically from the magicians data.
However, this complicates things when dealing with the magicianId in the magician-details view.
For now it's hardcode but once the magician-details is working, it would be best to make the magician names dynamic again.